refactor: Fixed some visual bugs and added permissions to limit the view of some sections

+++ apps/home/views/clients.py
@@ -4,6 +4,10 @@ from django.core.paginator import Paginator
 from django.db.models import Q
 
 
+def get_permission(request, permission_type, model='client'):
+    return request.user.has_perm(f'home.{permission_type}_{model}')
+
+
 def get_paginated_clients(request, order_by):
     # query = Q()
     # clients_list = Client.objects.filter(query)
@@ -21,6 +25,9 @@ def get_paginated_clients(request, order_by):
 
 
 def home(request, order_by=None):
+    if not get_permission(request, 'view'):
+        return render(request, 'home/page-404.html')
+
     paginator, clients = get_paginated_clients(request, order_by=order_by)
 
     context = {
@@ -32,6 +39,9 @@ def home(request, order_by=None):
 
 
 def create(request):
+    if not get_permission(request, 'add'):
+        return render(request, 'home/page-404.html')
+
     if request.method == 'POST':
         client = Client(
             name=request.POST.get('name', ''),
@@ -39,7 +49,7 @@ def create(request):
             email=request.POST.get('email', ''),
             phone=request.POST.get('phone', ''),
             area=request.POST.get('area', ''),
-            location=request.POST.get('location', ''),
+            location=request.POST.get('address', ''),
             description=request.POST.get('about', ''),
         )
 
@@ -48,7 +58,30 @@ def create(request):
     return redirect('clients_home')
 
 
+def edit(request, client_id):
+    if not get_permission(request, 'change'):
+        return render(request, 'home/page-404.html')
+
+    if request.method == 'POST':
+        client = Client.objects.get(id=client_id)
+
+        client.name = request.POST.get('name', client.name)
+        client.cnpj = request.POST.get('cnpj', client.cnpj)
+        client.email = request.POST.get('email', client.email)
+        client.phone = request.POST.get('phone', client.phone)
+        client.area = request.POST.get('area', client.area)
+        client.location = request.POST.get('address', client.location)
+        client.description = request.POST.get('about', client.description)
+
+        client.save()
+
+    return redirect('clients_home')
+
+
 def delete(request, client_id):
+    if not get_permission(request, 'delete'):
+        return render(request, 'home/page-404.html')
+
     client = Client.objects.get(id=client_id)
     client.delete()
